Best Linear Unbiased Predictions from dosresmeta Models

This method function computes (empirical) best linear unbiased predictions from fitted dose-response meta-analytical models represented in objects of class "dosresemta".

# S3 method for dosresmeta
blup(object, ...)

Arguments

object
objects of classe "dosresmeta".

…
further arguments passed to or from other methods.

Details

The method function blup produces (empirical) best linear unbiased predictions from dosresmeta objects. Predictions are expressed in terms of study-specific deviations as random effects. Predicted random effects from blup are a shrunk version of study-specific realizations, where study-specific predictions borrow strength from the assumption of an underlying distribution in a (usually hypothetical) population of studies. Blup are not avaialable for fixed-effects models since the are meaningless in that context.

Examples

## Load data and run the linear and quadratic models
data("alcohol_cvd")
lin <- dosresmeta(formula = logrr ~ dose, type = type, id = id,
                  se = se, cases = cases, n = n, data = alcohol_cvd)
quadr <- dosresmeta(formula = logrr ~ dose + I(dose^2), type = type, id = id,
                    se = se, cases = cases, n = n, data = alcohol_cvd)

## blup prediction for the previous models
blup(lin)

#>               dose
#> [1,]  0.0120899989
#> [2,] -0.0072113932
#> [3,] -0.0002240596
#> [4,]  0.0019372415
#> [5,] -0.0039281451
#> [6,] -0.0026636425

blup(quadr)

#>               dose     I(dose^2)
#> [1,]  0.0103422742 -1.893023e-04
#> [2,]  0.0029688170 -5.109482e-04
#> [3,] -0.0018167044  1.148953e-04
#> [4,]  0.0124878826 -4.263793e-04
#> [5,] -0.0241690913  9.734803e-04
#> [6,]  0.0001868219  3.825425e-05