Computes the covariance matrix for a set of log relative risks

Reconstructs the covariance matrix for a set of (reported) log relative risks, given the number of cases and the number of total persons or person-years for each treatment (dose) level.

covar.logrr(cases, n, y, v, type, data, covariance = "gl")

Arguments

cases
a vector, defining the number of cases for each exposure level.

n
a vector, defining the total number of subjects for each exposure level. For incidence-rate data n indicates the amount of person-time within each exposure level.

y
a vector, defining the (reported) log relative risks.

v
a vector, defining the variances of the reported log relative risks.

type
a vector (or a character string), specifying the design of the study. Options are cc, ir, and ci, for case-control, incidence-rate, and cumulative incidence data, respectively.

data
an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables in the previous arguments.

covariance
method to approximate the coviariance among set of reported log relative risks, "gl" for the method proposed by Greenland and Longnecker (default), "h" for the method proposed by Hamling.

Value

The (co)variance matrix of the log relative risks.

Details

This is an internal function called by dosresmeta to reconstruct the (co)variance matrix of the (adjusted) log relative risks. The function calls, depending on the choosen method, grl (default) or hamling to reconstruct the effective counts corresponding to the (adjusted) log relative risks as well as their standard errors. From these it computes the covariance matrix; analytical formulas can be found in the referenced article.

References

Orsini, N., Li, R., Wolk, A., Khudyakov, P., Spiegelman, D. (2012). Meta-analysis for linear and nonlinear dose-response relations: examples, an evaluation of approximations, and software. American journal of epidemiology, 175(1), 66-73.

See also

Examples

## Loading data
data("alcohol_cvd")

## Obtaining the (co)variance matrix of log RR for the first study (id = 1)
covar.logrr(y = logrr, v = I(se^2), cases = cases, n = n, type = type,
            data = subset(alcohol_cvd, id == 1))

#>            [,1]       [,2]       [,3]       [,4]
#> [1,] 0.04987985 0.01875906 0.01678737 0.01780975
#> [2,] 0.01875906 0.05463996 0.01736485 0.01842239
#> [3,] 0.01678737 0.01736485 0.14176254 0.01648609
#> [4,] 0.01780975 0.01842239 0.01648609 0.19714011

## Obtaining the (co)variance matrices of log RRfor all study
by(alcohol_cvd, alcohol_cvd$id, function(x)
   covar.logrr(y = logrr, v = I(se^2), cases = cases, n = n,
               type = type, data = x))

#> alcohol_cvd$id: 1
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] 0.04987985 0.01875906 0.01678737 0.01780975
#> [2,] 0.01875906 0.05463996 0.01736485 0.01842239
#> [3,] 0.01678737 0.01736485 0.14176254 0.01648609
#> [4,] 0.01780975 0.01842239 0.01648609 0.19714011
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 2
#>            [,1]       [,2]       [,3]
#> [1,] 0.04897601 0.02669775 0.02510026
#> [2,] 0.02669775 0.08854517 0.02425041
#> [3,] 0.02510026 0.02425041 0.14983031
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 3
#>            [,1]       [,2]       [,3]
#> [1,] 0.10058223 0.03667789 0.04203026
#> [2,] 0.03667789 0.17523766 0.03757552
#> [3,] 0.04203026 0.03757552 0.82646389
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 4
#>            [,1]       [,2]       [,3]
#> [1,] 0.11104463 0.03566026 0.03466437
#> [2,] 0.03566026 0.11449601 0.03293283
#> [3,] 0.03466437 0.03293283 0.22216648
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 5
#>            [,1]       [,2]       [,3]
#> [1,] 0.02079807 0.01122058 0.01347285
#> [2,] 0.01122058 0.05681004 0.01269044
#> [3,] 0.01347285 0.01269044 0.07501320
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 6
#>             [,1]        [,2]        [,3]
#> [1,] 0.024519546 0.004898219 0.004718234
#> [2,] 0.004898219 0.123263570 0.004371731
#> [3,] 0.004718234 0.004371731 0.173481760

## Restructuring the previous results in a list of matrices
do.call("list", by(alcohol_cvd, alcohol_cvd$id, function(x)
   covar.logrr(y = logrr, v = I(se^2), cases = cases, n = n, type = type,
               data = x)))

#> $`1`
#>            [,1]       [,2]       [,3]       [,4]
#> [1,] 0.04987985 0.01875906 0.01678737 0.01780975
#> [2,] 0.01875906 0.05463996 0.01736485 0.01842239
#> [3,] 0.01678737 0.01736485 0.14176254 0.01648609
#> [4,] 0.01780975 0.01842239 0.01648609 0.19714011
#> 
#> $`2`
#>            [,1]       [,2]       [,3]
#> [1,] 0.04897601 0.02669775 0.02510026
#> [2,] 0.02669775 0.08854517 0.02425041
#> [3,] 0.02510026 0.02425041 0.14983031
#> 
#> $`3`
#>            [,1]       [,2]       [,3]
#> [1,] 0.10058223 0.03667789 0.04203026
#> [2,] 0.03667789 0.17523766 0.03757552
#> [3,] 0.04203026 0.03757552 0.82646389
#> 
#> $`4`
#>            [,1]       [,2]       [,3]
#> [1,] 0.11104463 0.03566026 0.03466437
#> [2,] 0.03566026 0.11449601 0.03293283
#> [3,] 0.03466437 0.03293283 0.22216648
#> 
#> $`5`
#>            [,1]       [,2]       [,3]
#> [1,] 0.02079807 0.01122058 0.01347285
#> [2,] 0.01122058 0.05681004 0.01269044
#> [3,] 0.01347285 0.01269044 0.07501320
#> 
#> $`6`
#>             [,1]        [,2]        [,3]
#> [1,] 0.024519546 0.004898219 0.004718234
#> [2,] 0.004898219 0.123263570 0.004371731
#> [3,] 0.004718234 0.004371731 0.173481760
#>