Computes mean and standardized mean differences for continuous outcome with corresponding co(variance) matrix

This internal function computes mean and standardized mean of a continuous outcome with the corresponding variances. It also reconstructs the covariance matrix from the available data.

covar.smd(y, sd, n, measure = "md", method = "cohens", data)

Arguments

y
a vector defining the mean outcome for each treatment level.

sd
a vector defining the standard deviation of the outcome for each treatment level.

n
a vector defining the number of subjects for each treatment level.

measure
character string, indicating the measure to be calculated. Options are md and smd for mean difference and standardized mean difference, respectively.

method
character string indicating the method to be used. Options are cohens, hedges, and glass.

data
an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables in the previous arguments.

Value

A list containing the following

y	mean or standardized mean differences for each treatment level, included the referent one (0 by calculation).

v	variances corresponding to the mean or standardized mean differences for each treatment level, included the referent one (0 by calculation)

S	co(variance) matrix for the non-referent mean or standardized mean differences.

Details

This is an internal function called by dosresmeta to reconstruct the (co)variance matrix of the outcome variable. The function is expected to be extended and/or modified at every release of the package

References

Cooper, H., Hedges, L. V., & Valentine, J. C. (Eds.). (2009). The handbook of research synthesis and meta-analysis. Russell Sage Foundation.

See also

Examples

## Loading the data
data("ari")

## Obtaining standardized mean differences, variances, and (co)varinace 
## matrix for the first study (id = 1)
covar.smd(y, sd, n, measure = "smd", data = subset(ari, id == 1))

#> $y
#>           [,1]
#> [1,] 0.0000000
#> [2,] 0.1599767
#> [3,] 0.2893776
#> [4,] 0.3275973
#> attr(,"scaled:center")
#> [1] 5.3
#> 
#> $v
#> [1] 0.00000000 0.02266982 0.02311697 0.02255206
#> 
#> $S
#>            [,1]       [,2]       [,3]
#> [1,] 0.02266982 0.01182900 0.01183749
#> [2,] 0.01182900 0.02311697 0.01189637
#> [3,] 0.01183749 0.01189637 0.02255206
#> 

## Obtaining mean differences, variances, and (co)varinace matrices for the all the studies
cov.md <- by(ari, ari$id, function(x) covar.smd(y, sd, n, "md", data = x))

## Extracting mean differences
unlist(lapply(cov.md, function(x) x$y))

#>    11    12    13    14    21    22    23    24    31    32    33    41    42 
#>  0.00  2.93  5.30  6.00  0.00 12.71  9.40 12.11  0.00 12.60  8.50  0.00  9.50 
#>    43    51    52    53    54 
#>  8.90  0.00  9.60 10.10 14.40 

## Extracting variances for the mean differences
unlist(lapply(cov.md, function(x) x$v))

#>        11        12        13        14        21        22        23        24 
#>  0.000000  7.592563  7.715466  7.514985  0.000000 13.343918 13.343918 13.764477 
#>        31        32        33        41        42        43        51        52 
#>  0.000000 12.038243 11.977154  0.000000  8.562825  8.654240  0.000000 25.447194 
#>        53        54 
#> 25.447194 24.701057 

## List of the (co)variance matrices for the mean differences
lapply(cov.md, function(x) x$S)

#> $`1`
#>          [,1]     [,2]     [,3]
#> [1,] 7.592563 3.944189 3.944189
#> [2,] 3.944189 7.715466 3.944189
#> [3,] 3.944189 3.944189 7.514985
#> 
#> $`2`
#>           [,1]      [,2]      [,3]
#> [1,] 13.343918  6.366449  6.366449
#> [2,]  6.366449 13.343918  6.366449
#> [3,]  6.366449  6.366449 13.764477
#> 
#> $`3`
#>           [,1]      [,2]
#> [1,] 12.038243  5.779592
#> [2,]  5.779592 11.977154
#> 
#> $`4`
#>          [,1]     [,2]
#> [1,] 8.562825 4.338831
#> [2,] 4.338831 8.654240
#> 
#> $`5`
#>          [,1]     [,2]     [,3]
#> [1,] 25.44719 11.61461 11.61461
#> [2,] 11.61461 25.44719 11.61461
#> [3,] 11.61461 11.61461 24.70106
#>