Ancillary Parameters for Controlling the Fit in dosresmeta Models

This internal function sets the parameter options used for fitting dose-response meta-analytical models, commonly to pre-specified default values. It is usually internally called by dosresmeta.fit. The function has many more arguments than needed for dosresmeta. For more details see mixmeta.control.

dosresmeta.control(optim = list(), showiter = FALSE, maxiter = 100,
  initPsi = NULL, Psifix = NULL, Scor = NULL, addSlist = NULL,
  inputna = FALSE, inputvar = 10^4, loglik.iter = "hybrid",
  igls.inititer = 10, gr = FALSE, hessian = FALSE, vc.adj = TRUE,
  reltol = sqrt(.Machine$double.eps), checkPD = NULL,
  set.negeigen = sqrt(.Machine$double.eps))

Arguments

optim
list of parameters passed to the control argument of the function optim, which performs the quasi-Newton optimization in likelihood-based random-effects models. See optim.

showiter
logical. If TRUE, the progress of iterative optimization is shown.

maxiter
positive interger value. Maximum number of iterations in methods involving optimization procedures.

initPsi
either a matrix or a vector of its lower triangular elements (with diagonal, taken by column) from which starting values of the parameters of the between-study (co)variance matrix are derived, used in the optimization procedure for likelihood-based random-effects models. If NULL (the default, and recommended), the starting value is created internally through an iterative generalized least square algorithm.

Psifix
either a matrix or a vector of its lower triangular elements (with diagonal, taken by column), or optionally a named list with one or more of such objects. Used to define fixed parts of the random-effects

Scor
either a scalar, vector or matrix representing the within-unit correlation(s) to be inputted when the covariances are not provided in multivariate models, and ignored if they are.

addSlist
a list of m matrices for the (outer-level) groups of units defining the (known) error (co)variance structure, when this cannot be passed through the argument S of mixmeta.

inputna
logical. If missing values must be internally inputted. To be used with caution.

inputvar
multiplier for inputting the missing variances, to be passed as an argument to inputna.

loglik.iter
iterative scheme used in in likelihood-based optimization routines. Options are "hybrid", "newton", and "igls" or "RIGLS".

igls.inititer
number of iteration of the iterative generalized least square algorithm to be run in the hybrid optimization procedure of linkelihood-based models to provide the starting value.

gr
indicates if the gradient of the (re)ml likelihood should be provided. FALSE by default.

hessian
logical. If TRUE, the Hessian matrix of the parameters estimated in the optimization process is computed and returned. Only applicable to likelihood-based estimation methods.

vc.adj
logical. If TRUE, an adjustement to the way the marginal variance part is computed in the (co)variance components estimator is applied in the variance components estimator.

reltol
relative convergence tolerance in methods involving optimization procedures. The algorithm stops if it is unable to reduce the value by a factor of \(reltol * (abs(val) + reltol)\) at a step.

checkPD
logical. Determines if the semi-positiveness of within-unit error or random-effects (co)variance matrices must be checked.

set.negeigen
positive value. Value to which negative eigenvalues are to be set in estimators where such method is used to force semi-positive definiteness of the estimated between-study (co)variance matrix.

Value

A list with components named as the arguments.

References

Gasparrini, A., Armstrong, B., Kenward, M. G. (2012). Multivariate meta-analysis for non-linear and other multi-parameter associations. Statistics in Medicine, 31(29), 3821-3839.

See also

Examples

## Loading data
data("alcohol_cvd")

## print the iterations (see ?optim) and change the default for starting values
dosresmeta(formula = logrr ~ dose, type = type, id = id, se = se,
           cases = cases, n = n, data = alcohol_cvd, proc = "1stage",
           control = list(showiter = TRUE, igls.inititer = 20))

#> Warning: one-dimensional optimization by Nelder-Mead is unreliable:
#> use "Brent" or optimize() directly

#>   Nelder-Mead direct search function minimizer
#> function value for initial parameters = 18.647688
#>   Scaled convergence tolerance is 2.77872e-07
#> Stepsize computed as 0.000860
#> BUILD              2 18.647688 18.606595
#> REFLECTION         4 18.606595 18.601346
#> HI-REDUCTION       6 18.601346 18.600056
#> HI-REDUCTION       8 18.600056 18.599785
#> HI-REDUCTION      10 18.599785 18.599684
#> HI-REDUCTION      12 18.599684 18.599676
#> HI-REDUCTION      14 18.599676 18.599666
#> HI-REDUCTION      16 18.599667 18.599666
#> Exiting from Nelder Mead minimizer
#>     18 function evaluations used

#> Call:  dosresmeta(formula = logrr ~ dose, id = id, type = type, cases = cases, 
#>     n = n, data = alcohol_cvd, se = se, proc = "1stage", control = list(showiter = TRUE, 
#>         igls.inititer = 20))
#> 
#> Fixed-effects coefficients:
#>    dose  
#> -0.0044  
#> 
#> 6 studies, 19 values, 1 fixed and 1 random-effects parameters
#>   logLik       AIC       BIC  
#> -18.5997   41.1993   42.9801  
#>