ML and REML Estimators for dosresmeta Models

These functions implement maximum likeliihood (ML) and restricted maximum likelihood (REML) estimators for random-effects dose-response meta-analysis. They are meant to be used internally and not directly run by the users.

dosresmeta.ml(Xlist, Zlist, ylist, Slist, nalist, q, nall, control, ...)

dosresmeta.reml(Xlist, Zlist, ylist, Slist, nalist, q, nall, control, ...)

Arguments

Xlist
a m-dimensional list of study-specific design matrices for the fixed-effects part of the model.

Zlist
a m-dimensional list of study-specific design matrices for the random-effects part of the model.

ylist
a m-dimensional list of study-specific of vectors of estimated outcomes.

Slist
a m-dimensional list of within-study (co)variance matrices of estimated outcomes.

nalist
a m-dimensional list of k-dimensional study-specific logical vectors, identifying missing outcomes.

q
numeric scalars: number of predictors, number of observations (excluding missing).

nall
numeric scalars: number of predictors, number of observations (excluding missing).

control
list of parameters for controlling the fitting process, usually internally set to default values by dosresmeta.control.

...
further arguments passed to or from other methods. Currently not used.

Value

These functions return an intermediate list object, whose components are then processed by dosresmeta.fit. Other components are added later through dosresmeta to finalize an object of class "dosresmeta".

Details

The estimation involves \(p\) fixed-effects coefficients and the \(p(p+1)/2\) random-effects parameters defining the between-study (co)variance matrix. The hybrid estimation procedure is based first on few runs of iterative generalized least square algorithm and then quasi-Newton iterations, using specific likelihood functions, until convergence. The estimation algorithm adopts a profiled (or concentrated) approach, that is expressed only in terms of the random-effects parameters. Cholesky and and QR decompositions are used for computational stability and efficiency, and for assuring the positive-definiteness of the estimated between-study (co)variance matrix. See the help page for the likelihood functions for further details.

References

Gasparrini, A., Armstrong, B., Kenward, M. G. (2012). Multivariate meta-analysis for non-linear and other multi-parameter associations. Statistics in Medicine, 31(29), 3821-3839.

See also

Examples

data("alcohol_cvd")

## Random-effect dose-response model assuming linearity, ML estimator
lin.ml <- dosresmeta(formula = logrr ~ dose, type = type, id = id,
                     se = se, cases = cases, n = n, data = alcohol_cvd,
                     , method = "ml")
summary(lin.ml)

#> Call:  dosresmeta(formula = logrr ~ dose, id = id, type = type, cases = cases, 
#>     n = n, data = alcohol_cvd, se = se, method = "ml")
#> 
#> Two-stage random-effects meta-analysis
#> Estimation method: ML
#> Covariance approximation: Greenland & Longnecker
#> 
#> Chi2 model: X2 = 0.6610 (df = 1), p-value = 0.4162
#> 
#> Fixed-effects coefficients
#>              Estimate  Std. Error        z  Pr(>|z|)  95%ci.lb  95%ci.ub   
#> (Intercept)   -0.0043      0.0053  -0.8130    0.4162   -0.0148    0.0061   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1 
#> 
#> Between-study random-effects (co)variance components
#>   Std. Dev
#>     0.0086
#> 
#> Univariate Cochran Q-test for residual heterogeneity:
#> Q = 14.1464 (df = 5), p-value = 0.0147
#> I-square statistic = 64.7%
#> 
#> 6 studies, 6 values, 1 fixed and 1 random-effects parameters
#>   logLik       AIC       BIC  
#>  17.4051  -30.8102  -31.2267  
#> 

## Random-effect dose-response model assuming linearity, REML estimator
lin.reml <- dosresmeta(formula = logrr ~ dose, type = type, id = id,
                       se = se, cases = cases, n = n, data = alcohol_cvd,
                       , method = "reml")
summary(lin.reml)

#> Call:  dosresmeta(formula = logrr ~ dose, id = id, type = type, cases = cases, 
#>     n = n, data = alcohol_cvd, se = se, method = "reml")
#> 
#> Two-stage random-effects meta-analysis
#> Estimation method: REML
#> Covariance approximation: Greenland & Longnecker
#> 
#> Chi2 model: X2 = 0.5495 (df = 1), p-value = 0.4585
#> 
#> Fixed-effects coefficients
#>              Estimate  Std. Error        z  Pr(>|z|)  95%ci.lb  95%ci.ub   
#> (Intercept)   -0.0044      0.0059  -0.7413    0.4585   -0.0159    0.0072   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1 
#> 
#> Between-study random-effects (co)variance components
#>   Std. Dev
#>     0.0100
#> 
#> Univariate Cochran Q-test for residual heterogeneity:
#> Q = 14.1464 (df = 5), p-value = 0.0147
#> I-square statistic = 64.7%
#> 
#> 6 studies, 6 values, 1 fixed and 1 random-effects parameters
#>   logLik       AIC       BIC  
#>  14.0341  -24.0683  -24.8494  
#>