dosresmeta Object

An object returned by dosresmeta function, inheriting from class "dosresmeta", and representing a fitted dose-response (meta-analytical) model.

Value

Objects of class "dosresmeta" are lists with defined components. Dimensions of such components differs according to the choosen procedure. For the one-stage analysis the dimensions refer to a one dimensional outcome, \(p\) predictors and \(m\) studies used for fitting the model. For the two-stage analysis the dimensions refer to \(p\) outcome parameters, no predictor (only the intercept) and \(m\) studies. The following components needs to be included in a legitimate dosresmeta object:

coefficients	a \(p\)-dimensional vector of the fixed-effects coefficients.

vcov	estimated \(p \times p\) (co)variance matrix of the fixed-effects coefficients.

Psi	for random-effects models, the estimated \(p \times p\) between-study (co)variance matrix.

residuals	a vector of residuals, that is observed minus fitted values.

fitted.values	a vector of of fitted mean values.

df.residual	the residual degrees of freedom.

rank	the numeric rank of the fitted model.

logLik	the (restricted) log-likelihood of the fitted model.

converged,niter	for models with iterative estimation methods, logical scalar indicating if the algorithm eventually converged, and number of iterations.

control	a list with the values of the control arguments used, as returned by dosresmeta.control.

method	the estimation method.

dim	list with the following scalar components: \(m\) (number of studies included in estimation, \(k\) (number of outcome parameters), \(p\) (number of predictors).

df	list with the following scalar components: nall (number of observations used for estimation, excluding missing values), nobs (equal to nall, minus the number of fixed-effects coefficients in REML models), fixed (number of estimated fixed-effects coefficients), random (number of estimated (co)variance terms).

lab	list with the following label vectors: \(p\) for the p predictors (including intercept).

model	the model frame used for fitting.

call	the function call.

formula	the model supplied.

terms	the terms object representing the fitted model.

proc	the estimation procedure.

center	if the desing matrix had been centered.

covariance	how the (co)variance had been appproximated.

Slist	list of approximated (co)variance matrices.

id	identification vector of the studies.

v	variances of the outcome values

Methods

A number of methods functions are available for dosresmeta objects, most of them common to other regression functions. Specifically-written method functions are defined for predict (standard predictions). The qtest method performs the Cochran Q test for heterogeneity only for a two-stage analysis. Other methods have been produced for summary, logLik, coef, and vcov. Printing functions for the objects of classes defined above are also provided. All the methods above are visible (exported from the namespace) and documented. In additions, several default method functions for regression are also applicable to objects of class "mixmeta", such as fitted, residuals, AIC, BIC and update, among others.

See also