Approximating effective-counts as proposed by Greenland & Longnecker

Reconstructs the set of pseudo-numbers (or 'effective' numbers) of cases and non-cases consistent with the input data (log relative risks). The method was first proposed in 1992 by Greenland and Longnecker.

grl(y, v, cases, n, type, data, tol = 1e-05)

Arguments

y
a vector, defining the (reported) log relative risks.

v
a vector, defining the variances of the reported log relative risks.

cases
a vector, defining the number of cases for each exposure level.

n
a vector, defining the total number of subjects for each exposure level. For incidence-rate data n indicates the amount of person-time within each exposure level.

type
a vector (or a character string), specifying the design of the study. Options are cc, ir, and ci, for case-control, incidence-rate, and cumulative incidence data, respectively.

data
an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables in the previous arguments.

tol
define the tolerance.

Value

The results are returned structured in a matrix

A	approximated number of effective cases.

N	approximated total number of effective subjects.

Details

The function reconstructs the effective counts corresponding to the multivariable adjusted log relative risks as well as their standard errors. A unique solution is guaranteed by keeping the margins of the table of pseudo-counts equal to the margins of the crude or unadjusted data (Greenland and Longnecker 1992). See the referenced article for a complete description of the algorithm implementation.

References

Greenland, S., Longnecker, M. P. (1992). Methods for trend estimation from summarized dose-response data, with applications to meta-analysis. American journal of epidemiology, 135(11), 1301-1309.

Orsini, N., Li, R., Wolk, A., Khudyakov, P., Spiegelman, D. (2012). Meta-analysis for linear and nonlinear dose-response relations: examples, an evaluation of approximations, and software. American journal of epidemiology, 175(1), 66-73.

See also

Examples

## Loading data
data("alcohol_cvd")

## Obtaining pseudo-counts for the first study (id = 1)
grl(y = logrr, v = I(se^2), cases = cases, n = n, type = type,
   data = subset(alcohol_cvd, id == 1))

#>              A   N
#> [1,] 125.04575 414
#> [2,]  67.12135 261
#> [3,]  68.86106 228
#> [4,]  18.64972  44
#> [5,]  17.32196  34

## Obtaining pseudo-counts for all study
by(alcohol_cvd, alcohol_cvd$id, function(x)
   grl(y = logrr, v = I(se^2), cases = cases, n = n, type = type, data = x))

#> alcohol_cvd$id: 1
#>              A   N
#> [1,] 125.04575 414
#> [2,]  67.12135 261
#> [3,]  68.86106 228
#> [4,]  18.64972  44
#> [5,]  17.32196  34
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 2
#>             A   N
#> [1,] 76.32634 258
#> [2,] 88.59093 413
#> [3,] 25.24795 202
#> [4,] 11.83473  64
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 3
#>              A   N
#> [1,] 81.774216 208
#> [2,] 46.587219 175
#> [3,] 15.641699  58
#> [4,]  5.996615  11
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 4
#>             A   N
#> [1,] 75.74649 208
#> [2,] 35.15622 117
#> [3,] 39.59892  81
#> [4,] 11.49838  39
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 5
#>              A   N
#> [1,] 153.24400 480
#> [2,] 230.45840 975
#> [3,]  41.20934 207
#> [4,]  40.08832 133
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 6
#>               A    N
#> [1,] 343.064318 1269
#> [2,]  99.685219  436
#> [3,]  13.149854   49
#> [4,]   9.100567   41

## Restructuring the previous results in a matrix
do.call("rbind", by(alcohol_cvd, alcohol_cvd$id, function(x)
   grl(y = logrr, v = I(se^2), cases = cases, n = n, type = type, data = x)))

#>                A    N
#>  [1,] 125.045754  414
#>  [2,]  67.121348  261
#>  [3,]  68.861059  228
#>  [4,]  18.649722   44
#>  [5,]  17.321957   34
#>  [6,]  76.326336  258
#>  [7,]  88.590929  413
#>  [8,]  25.247951  202
#>  [9,]  11.834726   64
#> [10,]  81.774216  208
#> [11,]  46.587219  175
#> [12,]  15.641699   58
#> [13,]   5.996615   11
#> [14,]  75.746492  208
#> [15,]  35.156220  117
#> [16,]  39.598921   81
#> [17,]  11.498375   39
#> [18,] 153.244004  480
#> [19,] 230.458403  975
#> [20,]  41.209336  207
#> [21,]  40.088318  133
#> [22,] 343.064318 1269
#> [23,]  99.685219  436
#> [24,]  13.149854   49
#> [25,]   9.100567   41