Approximating effective-counts as proposed by Hamling

Reconstructs the set of pseudo-numbers (or "effective" numbers) of cases and non-cases consistent with the input data (log relative risks). The method was first proposed in 2008 by Hamling.

hamling(y, v, cases, n, type, data)

Arguments

y
a vector, defining the (reported) log relative risks.

v
a vector, defining the variances of the reported log relative risks.

cases
a vector, defining the number of cases for each exposure level.

n
a vector, defining the total number of subjects for each exposure level. For incidence-rate data n indicates the amount of person-time within each exposure level.

type
a vector (or a character string), specifying the design of the study. Options are cc, ir, and ci, for case-control, incidence-rate, and cumulative incidence data, respectively.

data
an optional data frame (or object coercible by as.data.frame to a data frame) containing the variables in the previous arguments.

Value

A list containing the following

y	mean or standardized mean differences for each treatment level, included the referent one (0 by calculation).

v	variances corresponding to the mean or standardized mean differences for each treatment level, included the referent one (0 by calculation)

S	co(variance) matrix for the non-referent mean or standardized mean differences.

Details

The function reconstructs the effective counts corresponding to the multivariable adjusted log relative risks as well as their standard errors. A unique solution is guaranteed by keeping the ratio non-cases to cases and the fraction of unexposed subjects equal to the unadjusted data (Hamling). See the referenced article for a complete description of the algorithm implementation.

References

Hamling, J., Lee, P., Weitkunat, R., Ambuhl, M. (2008). Facilitating meta-analyses by deriving relative effect and precision estimates for alternative comparisons from a set of estimates presented by exposure level or disease category. Statistics in medicine, 27(7), 954-970.

Orsini, N., Li, R., Wolk, A., Khudyakov, P., Spiegelman, D. (2012). Meta-analysis for linear and nonlinear dose-response relations: examples, an evaluation of approximations, and software. American journal of epidemiology, 175(1), 66-73.

See also

Examples

## Loading data
data("alcohol_cvd")

## Obtaining pseudo-counts for the first study (id = 1)
hamling(y = logrr, v = I(se^2), cases = cases, n = n, type = type,
data = subset(alcohol_cvd, id == 1))

#>             A         N
#> [1,] 77.22402 256.72053
#> [2,] 42.86376 167.40232
#> [3,] 39.59588 131.64021
#> [4,] 14.04864  33.25695
#> [5,] 11.37916  22.39970

## Obtaining pseudo-counts for all study
by(alcohol_cvd, alcohol_cvd$id, function(x)
hamling(y = logrr, v = I(se^2), cases = cases, n = n, type = type, data = x))

#> alcohol_cvd$id: 1
#>             A         N
#> [1,] 77.22402 256.72053
#> [2,] 42.86376 167.40232
#> [3,] 39.59588 131.64021
#> [4,] 14.04864  33.25695
#> [5,] 11.37916  22.39970
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 2
#>              A         N
#> [1,] 52.936594 177.74712
#> [2,] 57.793169 267.42554
#> [3,] 18.561968 147.28031
#> [4,]  9.998022  53.65116
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 3
#>              A          N
#> [1,] 41.767484 105.948046
#> [2,] 22.346167  83.663073
#> [3,] 10.101682  37.333962
#> [4,]  2.800649   5.126876
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 4
#>              A         N
#> [1,] 41.049928 113.43178
#> [2,] 19.560563  65.54781
#> [3,] 25.512473  52.44973
#> [4,]  7.685215  26.24831
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 5
#>              A         N
#> [1,]  83.91138 242.21950
#> [2,] 112.60335 492.48801
#> [3,]  22.27534 121.32119
#> [4,]  15.84890  49.72784
#> ------------------------------------------------------------ 
#> alcohol_cvd$id: 6
#>               A         N
#> [1,] 194.789854 711.65888
#> [2,]  51.584127 235.57631
#> [3,]   8.465262  31.23998
#> [4,]   5.940076  28.18427

## Restructuring the previous results in a matrix
do.call("rbind", by(alcohol_cvd, alcohol_cvd$id, function(x)
   hamling(y = logrr, v = I(se^2), cases = cases, n = n, type = type,
      data = x)))

#>                A          N
#>  [1,]  77.224016 256.720533
#>  [2,]  42.863757 167.402324
#>  [3,]  39.595879 131.640213
#>  [4,]  14.048644  33.256951
#>  [5,]  11.379162  22.399697
#>  [6,]  52.936594 177.747117
#>  [7,]  57.793169 267.425543
#>  [8,]  18.561968 147.280311
#>  [9,]   9.998022  53.651163
#> [10,]  41.767484 105.948046
#> [11,]  22.346167  83.663073
#> [12,]  10.101682  37.333962
#> [13,]   2.800649   5.126876
#> [14,]  41.049928 113.431784
#> [15,]  19.560563  65.547809
#> [16,]  25.512473  52.449734
#> [17,]   7.685215  26.248314
#> [18,]  83.911375 242.219498
#> [19,] 112.603350 492.488014
#> [20,]  22.275337 121.321192
#> [21,]  15.848900  49.727836
#> [22,] 194.789854 711.658884
#> [23,]  51.584127 235.576313
#> [24,]   8.465262  31.239980
#> [25,]   5.940076  28.184271