Likelihood Functions for dosresmeta Models

These functions compute the value of the log-likelihood for random-effects dose-response meta-analysis, in terms of model parameters. They are meant to be used internally and not directly run by the users.

remlprof.fn(par, Xlist, Zlist, ylist, Slist, nalist, q, nall, ctrl)

remlprof.gr(par, Xlist, ylist, Slist, nalist, p, nall, ctrl)

mlprof.fn(par, Xlist, Zlist, ylist, Slist, nalist, q, nall, ctrl)

mlprof.gr(par, Xlist, ylist, Slist, nalist, p, nall, ctrl)

iter.igls(Psi, Xlist, Zlist, ylist, Slist, nalist, q)

Arguments

par
a vector representing the random-effects parameters defining the between-study (co)variance matrix.

Xlist
a m-dimensional list of study-specific design matrices for the fixed-effects part of the model.

Zlist
a m-dimensional list of study-specific design matrices for the random-effects part of the model.

ylist
a m-dimensional list of study-specific of vectors of estimated outcomes.

Slist
a m-dimensional list of within-study (co)variance matrices of estimated outcomes.

nalist
a m-dimensional list of k-dimensional study-specific logical vectors, identifying missing outcomes.

ctrl
list of parameters for controlling the fitting process, usually internally set to default values by dosresmeta.control.

p, q, nall
numeric scalars: number of predictors, number of observations (excluding missing).

Psi
a p x p matrix representing the current estimate of the between-study (co)variance matrix.

Value

mlprof.fn and remlprof.fn return the value of the (restricted) log-likelihood for a given set of parameters in par. iter.igls returns an updated estimate of Psi given its initial value or the value at the previous iteration.

Details

These functions are called internally by the fitting functions dosresmeta.ml and dosresmeta.reml to perform iterative optimization algorithms for estimating random effects meta-analytical models.

The maximization of the (restricted) likelihood starts with few runs of an iterative generalized least square algorithm implemented in iter.igls. This can be regarded as a fast and stable way to get starting values close to the maximum for the Quasi-Newton iterative algorithm, implemented in optim. Alternatively, starting values can be provided by the user in the control list (see mixmeta.control).

These functions actually specify the profiled version of the (restricted) likelihood, expressed only in terms of random-effects parameters, while the estimate of the fixed-effects coefficients is provided at each iteration by the internal function glsfit, based on the current value of the between-study (co)variance matrix. At convergence, the value of this profiled version is identical to the full (restricted) likelihood. This approach is computationally efficient, as it reduces the number of parameters in the optimization routine.

The parameterization of the between-study (co)variance matrix ensures the positive-definiteness of the estimated matrix. A Cholesky decomposition is then performed on the marginal (co)variance matrix in order to re-express the problem as standard least square equations, an approach which speeds up the computation of matrix inverses and determinants. These equations are finally solved through a QR decomposition, which guarantees stability.

See also