Summarizing dosresmeta Models

Print and summary method functions for dose-response models represented in objects of class "dosresmeta".

# S3 method for dosresmeta
print(x, digits = 4, ...)

# S3 method for dosresmeta
summary(object, ci.level = 0.95, ...)

# S3 method for summary.dosresmeta
print(x, digits = max(3, getOption("digits")
  - 3), signif.stars = getOption("show.signif.stars"), ...)

Arguments

x
an object of class dosresmeta or summary.dosresmeta produced by dosresmeta or summary.dosresmeta, respectively.

digits
an integer specifying the number of digits to which printed results must be rounded.

…
further arguments passed to or from other methods.

object
an object of class dosresmeta produced by dosresmeta.

ci.level
the confidence level used for defining the confidence intervals for the estimates of the (fixed-effects) coefficients.

signif.stars
logical. If TRUE, 'significance stars' are printed for each coefficient.

Value

The summary method function for dosresmeta objects produces a list of class "summary.dosresmeta". The components of the lists are some of those stored in the related dosresmeta object, plus the following:

AIC	the value of the Akaike information criterion for the fitted dosresmeta model, obtained through a call to AIC.

BIC	the value of the Bayesian information criterion for the fitted dosresmeta model, obtained through a call to BIC

corFixed	the \(p \times p\) correlation matrix of the fixed-effects coefficients, obtained from the (co)variance matrix vcov

corRandom	the \(p \times p\) correlation matrix of the random effects, obtained from the between-study (co)variance matrix \(\Psi\)

qstat	results from the Cochran Q test for heterogeneity.

ci.level	the confidence level used for defining the confidence intervals for the estimates of the fixed-effects coefficients.

chisq	overall test similar to anova.

As usual, the print method functions for classes "dosresmeta" and "summary.dosresmeta" do not return any value.

Details

the print method for class dosresmeta only returns basic information of the fitted model, namely the call, estimated (fixed-effects) coefficients, dimensions, and fit statistics (log-likelihood, AIC, BIC).

The summary method function computes additional statistics and tests, and produces a list object of class summary.dosresmeta. The print method function for this class, depending on the number of studies included in the analysis, shows additional information, such as tables reporting the estimates for the fixed and random-effects parts of the model, Chi-square test for model significance, Cochran Q test for heterogeneity and I-square.

See also

Examples

## Load data and run the model
data("alcohol_cvd")
model <- dosresmeta(formula = logrr ~ dose + I(dose^2), type = type, id = id,
                    se = se, cases = cases, n = n, data = alcohol_cvd)
## Defult print
model

#> Call:  dosresmeta(formula = logrr ~ dose + I(dose^2), id = id, type = type, 
#>     cases = cases, n = n, data = alcohol_cvd, se = se)
#> 
#> Fixed-effects coefficients:
#>      dose  I(dose^2)  
#>   -0.0302     0.0008  
#> 
#> 6 studies, 12 values, 2 fixed and 3 random-effects parameters
#>   logLik       AIC       BIC  
#>  34.6967  -59.3935  -57.8805  
#> 

## Specify digits
print(model, digit = 2)

#> Call:  dosresmeta(formula = logrr ~ dose + I(dose^2), id = id, type = type, 
#>     cases = cases, n = n, data = alcohol_cvd, se = se)
#> 
#> Fixed-effects coefficients:
#>      dose  I(dose^2)  
#>     -0.03       0.00  
#> 
#> 6 studies, 12 values, 2 fixed and 3 random-effects parameters
#> logLik     AIC     BIC  
#>  34.70  -59.39  -57.88  
#> 

## summary with 90th confidence intervals
summary(model, ci.level = .8)

#> Call:  dosresmeta(formula = logrr ~ dose + I(dose^2), id = id, type = type, 
#>     cases = cases, n = n, data = alcohol_cvd, se = se)
#> 
#> Two-stage random-effects meta-analysis
#> Estimation method: REML
#> Covariance approximation: Greenland & Longnecker
#> 
#> Chi2 model: X2 = 6.3694 (df = 2), p-value = 0.0414
#> 
#> Fixed-effects coefficients
#>            Estimate  Std. Error        z  Pr(>|z|)  80%ci.lb  80%ci.ub   
#> dose        -0.0302      0.0132  -2.2943    0.0218   -0.0471   -0.0133  *
#> I(dose^2)    0.0008      0.0005   1.5383    0.1240    0.0001    0.0014   
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1 
#> 
#> Between-study random-effects (co)variance components
#>            Std. Dev     Corr
#> dose         0.0209     dose
#> I(dose^2)    0.0008  -0.9063
#> 
#> Univariate Cochran Q-test for residual heterogeneity:
#> Q = 35.1177 (df = 10), p-value = 0.0001
#> I-square statistic = 71.5%
#> 
#> 6 studies, 12 values, 2 fixed and 3 random-effects parameters
#>   logLik       AIC       BIC  
#>  34.6967  -59.3935  -57.8805  
#>